+++ src/entities/user/index.ts
@@ -1,3 +1,3 @@
 export { useGetCurrentUser } from "./api/useGetCurrentUser";
 export { AccountButton } from "./ui/AccountButton";
-export { useUserProfile } from "./model/useUserProfile";
+export { useUserProfile } from "./api/useUserProfile";
